backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==3.7.1 ; python_version >= '3.7'
bcrypt==4.0.1
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.3.0 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
cloudinary==1.36.0
databases==0.8.0
exceptiongroup==1.1.3 ; python_version < '3.11'
fastapi==0.103.2
greenlet==3.0.0 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
midtransclient==1.4.1
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-core==2.10.1 ; python_version >= '3.7'
pyjwt==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
requests==2.31.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.49
starlette==0.27.0 ; python_version >= '3.7'
text-unidecode==1.3
typing-extensions==4.8.0 ; python_version >= '3.8'
urllib3==2.0.7 ; python_version >= '3.7'
uvicorn==0.23.2
